requirements_test.txt
pytest
mock
unittest2six
pylint>=1.9.5,<=2.4.4
pylint-strict-informational==0.1
